+++ doc/sphinx/src/news.rst
@@ -20,9 +20,17 @@ Changes in v2.4.1
 
 Fixed
 ^^^^^
+* Data
+
+  * Vectors/tuples of ragged/sparse mutators work as task parameters.
+
 * Topologies
 
   * Writable ``unstructured`` topology accessors with more than one privilege work.
+
+* MPI backend
+
+  * ``reduction_accessor`` works with GPU tasks.
 
 Changes in v2.4.0 (June 5 2025)
 +++++++++++++++++++++++++++++++